Refactor MethylC track into split and combined components
Separated the MethylC visualization into SplitMethylC and CombinedMethylC components. Moved shared rendering logic to a new helpers.tsx file. Updated displayTrack to use the new components for split and combined display modes.
