+++ src/components/tracks/displayTrack.tsx
@@ -15,7 +15,8 @@ import Importance from "./importance/importance";
 import DenseMotif from "./motif/dense";
 import SquishMotif from "./motif/squish";
 import BulkBed from "./bulkbed/bulkbed";
-import MethylC from "./methylC/methylc";
+import SplitMethylC from "./methylC/split";
+import CombinedMethylC from "./methylC/combined";
 
 export default function DisplayTrack({ id }: { id: string }) {
   const track = useTrackStore((state) => state.getTrack(id));
@@ -75,8 +76,8 @@ export const trackComponents: Record<TrackType, Partial<Record<DisplayMode, Reac
     [DisplayMode.Full]: BulkBed,
   },
   [TrackType.MethylC]: {
-    [DisplayMode.Combined]: MethylC,
-    [DisplayMode.Split]: MethylC,
+    [DisplayMode.Combined]: CombinedMethylC,
+    [DisplayMode.Split]: SplitMethylC,
   },
 };
