+++ src/components/tracks/methylC/combined.tsx
@@ -1,0 +1,3 @@
+export default function CombinedMethylC() {
+  return <g></g>;
+}
